[FE] rename EmptyContainer
